Add oguri_animation
The relationship is as follows: oguri -> animations -> outputs. This
will allow a given animation to be drawn synchronized to any number of
outputs. Outputs which are not assigned to an animation are stored in
oguri_state.idle_outputs. This way, outputs can be reassigned at
runtime, and animations can be added and removed as well.

I'm still not sure how I want to handle outputs which are configured but
do not currently exist.

Right now there's still only a single animation, and a single output
assigned to it.