Cache all resized frames as buffers
This prevents us from doing the check on buffer->busy that we really
should do so we don't step on the compositor during the first cycle when
we're actually modifying the buffers. After that, they're effectively
read-only and we can go back to ignoring it though.

With this change, we're down to 0.3% CPU with my usual test image! :D