Remove asserts in oguri_output_create
Instead, simply don't create an output if any part of it fails. I've
seen a number of crashes in here which I think are due to
swaywm/wlroots#1354, but better to handle it gracefully anyway.

This also makes it safe to call oguri_output_destroy at any stage of
creation.