Actually sync animations across outputs correctly
Turns out my original hack for tracking the first cycle of the animation
never worked correctly for all images, which means with some images we
were dropping frames once we started caching. I poked around the docs
again and found this alternate method which seems to work in all cases.
It's still sort of a misuse of the API, but what can you do?