Fix static images with buffer pool shrinking
This is technically the wrong thing to do, because we always want to
perform double buffering. I'm still pondering the best way to do this in
the case of only having a single frame. This also correctly schedules
a frame as soon as a new output is attached to an animation.